[K3 CP] Phase 0: expanded MHA by default, fused K concat, FlashKDA fixes, single-launch zigzag
Measured on 8x B200 with the 8-layer Kimi-K3-NVFP4 proxy (single 32K/64K prefill):
CP was 45-86% slower than TP8 because the CP dispatcher pins MLA layers to the
absorbed FA4 kernel (3.4x FLOPs, 64-row tiles: 11x slower per MLA layer than
expanded MHA). With these changes CP2/CP4/CP8 are within 1-2.5% of TP8.

- environ.py: SGLANG_K3_CP_MHA_MAX_WORKSPACE_MB default 0 -> 16384 MiB so K3 CP
  prefill uses expanded MHA without an opt-in (covers ~128K tokens at CP4/CP8).
  Behavior change of an existing knob; the shared MLA dispatcher is untouched.
- forward_mha.py: _concat_and_cast_mha_k no longer requires a power-of-two head
  count. The Triton wrapper already pads non-power-of-two heads, and K3 has
  12/24/48/96 heads per rank; the plain-copy fallback cost 1.2 ms per MLA layer
  at CP8/32K (2.4 ms at 64K) vs 0.62 ms for the kernel. Exact match verified for
  12/24/48/96 heads at 4K/64K tokens.
- cache_ops.py: both concat kernels index tokens in int64 (pid * k_stride0
  overflows int32 past ~116K tokens with 96 heads). Verified at 131072 x 96.
- kda_flashkda.py: FlashKDAKernel.extend returned an (out, None) tuple where
  KDAAttnBackend.forward_extend expects a tensor (crash in o_norm); return the
  tensor. New SGLANG_FLASHKDA_MAX_SEQ_LEN (default 2048, unchanged behavior)
  replaces the hard-coded Triton fallback threshold so long prefills can stay on
  the fused kernel.
- zigzag.py / flashattention_backend.py: opt-in SGLANG_CP_ZIGZAG_SINGLE_LAUNCH
  runs a CP rank's prev/next zigzag query blocks as one FA4 varlen launch on the
  expanded-MHA path (combined cu_seqlens_q/kv_len metadata; K base offsets
  repeated per request since FA4 reads cu_seqlens_k[i] only as the K base offset
  when seqused_k is given). Default off pending the A/B measurement.

+++ python/sglang/kernels/ops/kvcache/cache_ops.py
@@ -17,7 +17,8 @@ def concat_and_cast_mha_k_kernel(
     nope_dim: tl.constexpr,
     rope_dim: tl.constexpr,
 ):
-    pid_loc = tl.program_id(0)
+    # int64: pid_loc * k_stride0 exceeds int32 past ~116K tokens with 96 heads.
+    pid_loc = tl.program_id(0).to(tl.int64)
     head_range = tl.arange(0, head_cnt)
 
     k_head_ptr = k_ptr + pid_loc * k_stride0 + head_range[:, None] * k_stride1
@@ -62,7 +63,8 @@ def concat_and_cast_mha_k_pad_kernel(
     tl.arange needs a power-of-two extent, so walk HEAD_BLOCK == next_power_of_2(head_cnt)
     lanes and mask the tail.
     """
-    pid_loc = tl.program_id(0)
+    # int64: pid_loc * k_stride0 exceeds int32 past ~116K tokens with 96 heads.
+    pid_loc = tl.program_id(0).to(tl.int64)
     head_range = tl.arange(0, HEAD_BLOCK)
     head_mask = head_range < head_cnt
